+++ docs/CHANGELOG.rst
@@ -42,6 +42,7 @@ Added
 - Add JBrowse index output to ``hmr`` process
 - Add ``bamclipper`` tool and ``parallel`` package to ``resolwebio/dnaseq:4.2.0`` image
 - Support ``hg19_mm10`` hybrid genome in ``bam-split`` process
+- Support mappability-based normalization (RPKUM) in featureCounts
 
 Fixed
 -----